Renames kbn-eslint-import-resolver-kibana to osd-eslint-import-resolver-opensearch-dashboards (#37)
