+++ packages/osd-eslint-import-resolver-opensearch-dashboards/import_resolver_opensearch_dashboards.js
@@ -23,7 +23,7 @@ const webpackResolver = require('eslint-import-resolver-webpack');
 const nodeResolver = require('eslint-import-resolver-node');
 
 const {
-  getKibanaPath,
+  getOpenSearchDashboardsPath,
   getProjectRoot,
   getWebpackConfig,
   isFile,
@@ -39,8 +39,8 @@ function initContext(file, config) {
   }
 
   const projectRoot = getProjectRoot(file, config);
-  const kibanaPath = getKibanaPath(config, projectRoot);
-  const webpackConfig = getWebpackConfig(kibanaPath, projectRoot, config);
+  const opensearchDashboardsPath = getOpenSearchDashboardsPath(config, projectRoot);
+  const webpackConfig = getWebpackConfig(opensearchDashboardsPath, projectRoot, config);
   const aliasEntries = Object.entries(webpackConfig.resolve.alias || {});
 
   context = {
@@ -65,7 +65,7 @@ function tryNodeResolver(importRequest, file, config) {
   );
 }
 
-exports.resolve = function resolveKibanaPath(importRequest, file, config) {
+exports.resolve = function resolveOpenSearchDashboardsPath(importRequest, file, config) {
   config = config || {};
 
   if (config.forceNode) {
