+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/debug.js
@@ -17,12 +17,4 @@
  * under the License.
  */
 
-module.exports = Object.assign(
-  {},
-  require('./get_kibana_path'),
-  require('./get_project_root'),
-  require('./get_webpack_config'),
-  require('./get_path_type'),
-  require('./get_is_path_request'),
-  require('./resolve_webpack_alias')
-);
+exports.debug = require('debug')('eslint-plugin-import:resolver:opensearchDashboards');
