+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/get_is_path_request.js
