+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/get_path_type.js
@@ -23,7 +23,7 @@ const LRU = require('lru-cache');
 
 const DIR = Symbol('dir');
 const FILE = Symbol('file');
-const cache = process.env.KIBANA_RESOLVER_HARD_CACHE ? new Map() : new LRU({ maxAge: 1000 });
+const cache = process.env.OPENSEARCH_DASHBOARDS_RESOLVER_HARD_CACHE ? new Map() : new LRU({ maxAge: 1000 });
 
 function getPathType(path) {
   const cached = cache.get(path);
