+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/get_project_root.js
