+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/get_webpack_config.js
@@ -17,4 +17,21 @@
  * under the License.
  */
 
-exports.debug = require('debug')('eslint-plugin-import:resolver:kibana');
+const { resolve } = require('path');
+
+exports.getWebpackConfig = function (opensearchDashboardsPath) {
+  return {
+    context: opensearchDashboardsPath,
+    resolve: {
+      extensions: ['.js', '.json', '.ts', '.tsx'],
+      mainFields: ['browser', 'main'],
+      modules: ['node_modules', resolve(opensearchDashboardsPath, 'node_modules')],
+      alias: {
+        // Dev defaults for test bundle https://github.com/elastic/kibana/blob/6998f074542e8c7b32955db159d15661aca253d7/src/core_plugins/tests_bundle/index.js#L73-L78
+        fixtures: resolve(opensearchDashboardsPath, 'src/fixtures'),
+        test_utils: resolve(opensearchDashboardsPath, 'src/test_utils/public'),
+      },
+      unsafeCache: true,
+    },
+  };
+};
