+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/index.js
@@ -17,21 +17,12 @@
  * under the License.
  */
 
-const { resolve } = require('path');
-
-exports.getWebpackConfig = function (kibanaPath) {
-  return {
-    context: kibanaPath,
-    resolve: {
-      extensions: ['.js', '.json', '.ts', '.tsx'],
-      mainFields: ['browser', 'main'],
-      modules: ['node_modules', resolve(kibanaPath, 'node_modules')],
-      alias: {
-        // Dev defaults for test bundle https://github.com/elastic/kibana/blob/6998f074542e8c7b32955db159d15661aca253d7/src/core_plugins/tests_bundle/index.js#L73-L78
-        fixtures: resolve(kibanaPath, 'src/fixtures'),
-        test_utils: resolve(kibanaPath, 'src/test_utils/public'),
-      },
-      unsafeCache: true,
-    },
-  };
-};
+module.exports = Object.assign(
+  {},
+  require('./get_opensearch_dashboards_path'),
+  require('./get_project_root'),
+  require('./get_webpack_config'),
+  require('./get_path_type'),
+  require('./get_is_path_request'),
+  require('./resolve_webpack_alias')
+);
