+++ packages/osd-eslint-import-resolver-opensearch-dashboards/lib/resolve_webpack_alias.js
