+++ packages/osd-eslint-import-resolver-opensearch-dashboards/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@kbn/eslint-import-resolver-kibana",
-  "description": "eslint-plugin-import resolver for Kibana",
+  "name": "@osd/eslint-import-resolver-opensearch-dashboards",
+  "description": "eslint-plugin-import resolver for OpenSearch Dashboards",
   "private": true,
   "version": "2.0.0",
-  "main": "import_resolver_kibana.js",
+  "main": "import_resolver_opensearch_dashboards.js",
   "license": "Apache-2.0",
-  "kibana": {
+  "opensearchDashboards": {
     "devOnly": true
   },
   "repository": {
